requirements.txt
streamlit
yfinance
feedparser
beautifulsoup4
requests
openai
python-dotenv
pandas
plotly
google-genai
